make use of multi-inheritance to make units smaller and easier to understand
