+++ .gitignore
@@ -4,3 +4,6 @@
 build/
 dist/
 *.egg-info/
+.coverage
+.noseids
+coverage.xml
