+++ pyfritzhome/device/__init__.py
@@ -1,0 +1,15 @@
+from .alarm import FritzhomeAlarmDevice
+from .powermeter import FritzhomePowermeterDevice
+from .repeater import FritzhomeRepeaterDevice
+from .switch import FritzhomeSwitchDevice
+from .temperature import FritzhomeTemperatureDevice
+from .thermostat import FritzhomeThermostatDevice
+
+__all__ = [
+    "FritzhomeAlarmDevice",
+    "FritzhomePowermeterDevice",
+    "FritzhomeRepeaterDevice",
+    "FritzhomeSwitchDevice",
+    "FritzhomeTemperatureDevice",
+    "FritzhomeThermostatDevice",
+]
