+++ setup.cfg
@@ -1,2 +1,8 @@
 [flake8]
 max-line-length = 88
+
+[nosetests]
+with-coverage=1
+cover-xml=1
+cover-branches=1
+cover-package=pyfritzhome
