+++ tests/test_fritzhomepowermeterdevice.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from .abstractdevicetest import AbstractTestDevice
+
+
+class TestPowermeterDevice(AbstractTestDevice):
+    pass
